+++ custom_addons/odoo_ai_ops/tests/__init__.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-from . import test_order_risk, test_reconciliation
+from . import test_order_risk, test_reconciliation, test_shopify_client