+++ terraform/modules/ecs/cluster.tf
@@ -51,10 +51,10 @@ resource "aws_ecs_cluster_capacity_providers" "odoo" {
 
 resource "aws_cloudwatch_log_group" "odoo_logs" {
   name              = "/ecs/odoo"
-  retention_in_days = 7
+  retention_in_days = 14
 }
 
 resource "aws_cloudwatch_log_group" "fastapi_logs" {
   name              = "/ecs/fastapi"
-  retention_in_days = 7
+  retention_in_days = 14
 }