+++ terraform/modules/telemetry/langfuse.tf
@@ -2,7 +2,7 @@
 
 resource "aws_cloudwatch_log_group" "langfuse" {
   name              = "/ecs/langfuse"
-  retention_in_days = 7
+  retention_in_days = 14
 }
 
 resource "aws_ecs_task_definition" "langfuse" {