+++ terraform/outputs.tf
@@ -42,3 +42,16 @@ output "api_gateway_webhook_url" {
 output "webhook_queue_url" {
   value = module.webhooks.queue_url
 }
+
+# --- Invariants surfaced for `terraform test` (known at plan time) ---
+output "odoo_instance_type" { value = module.ecs.odoo_instance_type }
+output "pgbouncer_instance_type" { value = module.ecs.pgbouncer_instance_type }
+output "fastapi_instance_type" { value = module.ecs.fastapi_instance_type }
+output "clickhouse_instance_type" { value = module.ecs.clickhouse_instance_type }
+output "rds_instance_class" { value = module.data.db_instance_class }
+output "rds_multi_az" { value = module.data.db_multi_az }
+output "lambda_runtime" { value = module.webhooks.lambda_runtime }
+output "webhook_dlq_arn" { value = module.webhooks.dlq_arn }
+
+# SNS topic that all CloudWatch alarms publish to - subscribe here to get paged.
+output "alerts_topic_arn" { value = module.observability.sns_topic_arn }